--- src/components/InvalidImageModal.js
@@ -1,64 +1,0 @@
-import React from "react";
-import { Modal, Pressable, StyleSheet, Text, View } from "react-native";
-
-export default function InvalidImageModal({ modalText, setModalText }) {
-  return (
-    <Modal
-      animationType="slide"
-      transparent
-      visible={modalText !== ""}
-      onRequestClose={() => {
-        setModalText("");
-      }}
-    >
-      <View style={styles.centeredView}>
-        <View style={styles.modalView}>
-          <Text style={styles.modalText}>Not enough {modalText} detected. Please use a different photo.</Text>
-          <Pressable style={styles.modalButton} onPress={() => setModalText("")}>
-            <Text style={styles.modalButtonText}>Close</Text>
-          </Pressable>
-        </View>
-      </View>
-    </Modal>
-  )
-}
-
-const styles = StyleSheet.create({
-  centeredView: {
-    justifyContent: "center",
-    alignItems: "center",
-    marginTop: 240
-  },
-  modalView: {
-    margin: 20,
-    backgroundColor: "white",
-    borderRadius: 20,
-    padding: 25,
-    alignItems: "center",
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 2
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 4,
-    elevation: 5
-  },
-  modalButton: {
-    borderRadius: 20,
-    padding: 10,
-    elevation: 2,
-    backgroundColor: "#2196F3"
-  },
-  modalButtonText: {
-    color: "white",
-    fontWeight: "bold",
-    textAlign: "center"
-  },
-  modalText: {
-    marginBottom: 15,
-    textAlign: "center",
-    fontWeight: "bold",
-    color: "red"
-  }
-});
